requirements.txt
pyaml==20.4.0
pytest==5.4.3
faker==4.1.1